Add metric object to impression
From the docs: This object is associated with an impression as an array of metrics. These metrics can offer insight into
the impression to assist with decisioning such as average recent viewability, click-through rate, etc. Each
metric is identified by its type, reports the value of the metric, and optionally identifies the source or
vendor measuring the value.
We need it for adx, but i guess we can add it to all SSP since it's standart openrtb